refactor(trusty-search): split commands/discover.rs into submodules under 500-line cap (closes #473)
Convert the 559-line commands/discover.rs into a discover/ module with three
focused submodules, each well under the 500-line cap:

- discover/marker.rs  (244 lines): ProjectMarker enum, TRUSTY_TOOLS_DIR
  constant, detect_project_marker(), default_scan_paths(), and all
  detection-related unit tests (13 tests from the original file).
- discover/http.rs    (74 lines):  wait_for_daemon_ready() and
  fetch_known_index_ids() — the two async HTTP helper functions.
- discover/mod.rs     (288 lines): module-level doc comment, use declarations,
  auto_discover_and_index() orchestration, and the dead-root predicate tests.

Public API is identical: crate::commands::discover::auto_discover_and_index()
is the only external call site (start.rs:1390) and its path is unchanged.
mod.rs declares `mod http` and `mod marker` as private; all symbols from those
modules consumed by the orchestrator are re-imported via `use`.
